tegra: add smarter hotplug routine
This hotplug driver for cpus is based upon
@franciscofranco awesome work for various nexus
devices.

While i liked the idea to bound cpu0 and cpu2
as well as cpu1 and cpu3 together, i wanted to
let the hotplug driver be more on some sort of
high performance mode. In general my additions
should online 4 cores more often than the orginal
implementation and it should stay at this level
for some time until it recalculates the load.

Next (planned) things;
 * consider hotplug only for lowest thread intensive core
 * add thermal routines

Signed-off-by: Blechd0se <alex.christ@hotmail.de>