hotplug: consider current freq in for cpu_up()
When we decide to online one core, we should
also consider its current parent core. It may
happen that one core could be onlined while
the parent core is on a very low frequency and
just has a high load due that.

Also enable the full frequency table with the
last step (1300 mhz).

Signed-off-by: Blechd0se <alex.christ@hotmail.de>