hotplug: replace ktime with jiffies
Replacing ktime_get() with jiffies actually results in
much better timer handling. When we online a core we let
it stay online for at least 4 (4 * HZ) seconds.
During this interval we rearm our work_queue and let
the calculation happen again. In many case the algorithm
decides to not offline this core because when there is
some sort of average load happening.
Apart from a small performance increase this also could
help for battery life.

Also ensure the work_queue runs strictly on the first cpu.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>