hotplug: take average load of all cores in account
Sometimes it seems like we need to online a core only
because on one core is a really high load. It won't help
much in this case, since we might have just one process
eating up all ressources of this single core.

By taking the average load of all current onlined cores
into account, we can fix this issue a bit.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>